Move initialization functions into `esp-hal-common` package's `soc` module
