+++ esp-hal-common/src/soc/esp32c2/mod.rs
@@ -5,6 +5,9 @@
 //! The `SOC` module provides access, functions and structures that are useful
 //! for interacting with various system-related peripherals on `ESP32-C2` chip.
 
+use self::peripherals::{RTC_CNTL, TIMG0};
+use crate::{timer::Wdt, Rtc};
+
 pub mod efuse;
 pub mod gpio;
 pub mod peripherals;
@@ -17,4 +20,14 @@ pub(crate) mod registers {
 pub(crate) mod constants {
     pub const SOC_DRAM_LOW: u32 = 0x3FCA_0000;
     pub const SOC_DRAM_HIGH: u32 = 0x3FCE_0000;
+}
+
+#[export_name = "__post_init"]
+unsafe fn post_init() {
+    // RTC domain must be enabled before we try to disable
+    let mut rtc = Rtc::new(RTC_CNTL::steal());
+    rtc.swd.disable();
+    rtc.rwdt.disable();
+
+    Wdt::<TIMG0>::set_wdt_enabled(false);
 }
