+++ esp-hal-common/src/soc/esp32c3/mod.rs
@@ -9,6 +9,9 @@
 //!    * I2S_SCLK: 160_000_000 - I2S clock frequency
 //!    * I2S_DEFAULT_CLK_SRC: 2 - I2S clock source
 
+use self::peripherals::{RTC_CNTL, TIMG0, TIMG1};
+use crate::{timer::Wdt, Rtc};
+
 pub mod efuse;
 pub mod gpio;
 pub mod peripherals;
@@ -29,4 +32,15 @@ pub(crate) mod constants {
 
     pub const SOC_DRAM_LOW: u32 = 0x3FC8_0000;
     pub const SOC_DRAM_HIGH: u32 = 0x3FCE_0000;
+}
+
+#[export_name = "__post_init"]
+unsafe fn post_init() {
+    // RTC domain must be enabled before we try to disable
+    let mut rtc = Rtc::new(RTC_CNTL::steal());
+    rtc.swd.disable();
+    rtc.rwdt.disable();
+
+    Wdt::<TIMG0>::set_wdt_enabled(false);
+    Wdt::<TIMG1>::set_wdt_enabled(false);
 }
