+++ esp32s2-hal/src/lib.rs
@@ -60,62 +60,3 @@
 #![doc(html_logo_url = "https://avatars.githubusercontent.com/u/46717278")]
 
 pub use esp_hal_common::*;
-
-/// Function initializes ESP32 specific memories (RTC slow and fast) and
-/// then calls original Reset function
-///
-/// ENTRY point is defined in memory.x
-/// *Note: the pre_init function is called in the original reset handler
-/// after the initializations done in this function*
-#[cfg(feature = "rt")]
-#[doc(hidden)]
-#[no_mangle]
-pub unsafe extern "C" fn ESP32Reset() -> ! {
-    // These symbols come from `memory.x`
-    extern "C" {
-        static mut _rtc_fast_bss_start: u32;
-        static mut _rtc_fast_bss_end: u32;
-
-        static mut _rtc_slow_bss_start: u32;
-        static mut _rtc_slow_bss_end: u32;
-
-        static mut _stack_start_cpu0: u32;
-    }
-
-    // set stack pointer to end of memory: no need to retain stack up to this point
-    esp_hal_common::xtensa_lx::set_stack_pointer(&mut _stack_start_cpu0);
-
-    // copying data from flash to various data segments is done by the bootloader
-    // initialization to zero needs to be done by the application
-
-    // Initialize RTC RAM
-    esp_hal_common::xtensa_lx_rt::zero_bss(&mut _rtc_fast_bss_start, &mut _rtc_fast_bss_end);
-    esp_hal_common::xtensa_lx_rt::zero_bss(&mut _rtc_slow_bss_start, &mut _rtc_slow_bss_end);
-
-    // continue with default reset handler
-    esp_hal_common::xtensa_lx_rt::Reset();
-}
-
-/// The ESP32 has a first stage bootloader that handles loading program data
-/// into the right place therefore we skip loading it again.
-#[doc(hidden)]
-#[no_mangle]
-#[rustfmt::skip]
-pub extern "Rust" fn __init_data() -> bool {
-    false
-}
-
-#[export_name = "__post_init"]
-unsafe fn post_init() {
-    use esp_hal_common::{
-        peripherals::{RTC_CNTL, TIMG0, TIMG1},
-        timer::Wdt,
-    };
-
-    // RTC domain must be enabled before we try to disable
-    let mut rtc = Rtc::new(RTC_CNTL::steal());
-    rtc.rwdt.disable();
-
-    Wdt::<TIMG0>::set_wdt_enabled(false);
-    Wdt::<TIMG1>::set_wdt_enabled(false);
-}
